Squid proxy custom block form
Custom HTML/CSS form made to show up when users are blocked from accessing specific URL. It allows user to send a e-mail to network administrator with explanation why they think the page should not be blocked. The form collects user header data such as IP address, AD username and Squidguard group they are assigned to and appends it to the mail sent so the administrator can know where the message arived from.